Use decimal for money instead of double
Charge() and the balance getters used double, and callers converted the
decimal Request.Amount via Convert.ToDouble. Switch IPaymentService and
StripePaymentServiceProvider to decimal end-to-end, convert to Stripe's
integer-cent longs only at the boundary (AwayFromZero rounding), make the
balance getters divide by 100m (no lossy integer division), and update
PayoutModel.Balance/PendingBalance to decimal. Callers now pass
request.Amount directly.

Closes #22

+++ src/comissions.app.api/Models/PayoutModel.cs
@@ -4,8 +4,8 @@ public class PayoutModel
 {
     public int PayoutDelayDays { get; set; }
     public string Interval { get; set; }
-    public double Balance { get; set; }
+    public decimal Balance { get; set; }
     public bool Enabled { get; set; }
     public string PayoutUrl { get; set; }
-    public double PendingBalance { get; set; }
+    public decimal PendingBalance { get; set; }
 }
